Add addrgen tool
This is intended to generate object files to pass to the linker that
contain symbols for functions and other objects belonging to the game.

It takes a CSV file containing addresses, and a list of other object
files to be linked (so that it can tell which symbols are actually
used) as its input.